refactor: 💡 user 전역 상태 파일 이동
